Refactoring to get HTTP information

+++ inc/Http/HttpResponse.h
@@ -14,19 +14,19 @@ public:
     String body;
 
     static HttpResponse* createResponse();
-    static HttpResponse* createResponseWithHttpInformation(const String& httpVersion, uint8_t statusCode, const String& statusMessage);
+    static HttpResponse* createResponseWithHttpInformation(const String& httpVersion, long statusCode, const String& statusMessage);
 
     HttpResponse() { };
     virtual ~HttpResponse() { }
 
-    inline uint8_t getStatusCode() { return this->statusCode; }
+    inline long getStatusCode() { return this->statusCode; }
     inline const String& getHttpVersion() { return this->httpVersion; }
     inline const String& getStatusMessage() { return this->statusMessage; }
 
-    void setHttpInformation(const String& httpVersion, uint8_t statusCode, const String& statusMessage);
+    void setHttpInformation(const String& httpVersion, long statusCode, const String& statusMessage);
 
 protected:
-    uint8_t statusCode = 0;
+    long statusCode = 0;
     String httpVersion = "";
     String statusMessage = "";
