+++ src/Http/HttpResponse.cpp
@@ -7,7 +7,7 @@ APIPOLLER::HttpResponse *APIPOLLER::HttpResponse::createResponse()
 }
 
 
-APIPOLLER::HttpResponse *APIPOLLER::HttpResponse::createResponseWithHttpInformation(const String& httpVersion, uint8_t statusCode, const String& statusMessage)
+APIPOLLER::HttpResponse *APIPOLLER::HttpResponse::createResponseWithHttpInformation(const String& httpVersion, long statusCode, const String& statusMessage)
 {
     HttpResponse* responseWithHttpInformation = createResponse();
     responseWithHttpInformation->setHttpInformation(httpVersion, statusCode, statusMessage);
@@ -15,7 +15,7 @@ APIPOLLER::HttpResponse *APIPOLLER::HttpResponse::createResponseWithHttpInformat
 }
 
 
-void APIPOLLER::HttpResponse::setHttpInformation(const String& httpVersion, uint8_t statusCode, const String& statusMessage)
+void APIPOLLER::HttpResponse::setHttpInformation(const String& httpVersion, long statusCode, const String& statusMessage)
 {
     this->statusCode = statusCode;
     this->httpVersion = httpVersion;
